+++ podaac/subsetter/dimension_cleanup.py
@@ -11,11 +11,14 @@ dimension_cleanup.py
 ======================
 Functions which improve upon existing netCDF4 library existing functions
 """
-
 import collections
+from typing import List, Tuple
+
+import netCDF4 as nc
+import xarray as xr
 
 
-def remove_duplicate_dims(nc_dataset):
+def remove_duplicate_dims(nc_dataset: nc.Dataset) -> Tuple[nc.Dataset, List[str]]:
     """
     xarray cannot read netCDF4 datasets with duplicate dimensions.
     Function goes through a dataset to catch any variables with duplicate dimensions.
@@ -67,7 +70,7 @@ def remove_duplicate_dims(nc_dataset):
     return nc_dataset, dup_new_varnames
 
 
-def rename_dup_vars(dataset, rename_vars):
+def rename_dup_vars(dataset: xr.Dataset, rename_vars: List[str]) -> xr.Dataset:
     """
     NetCDF4 rename function raises and HDF error for variable in S5P files with duplicate dimensions
     This method will use xarray to rename the variables