+++ podaac/subsetter/subset.py
@@ -23,40 +23,44 @@ import functools
 import json
 import operator
 import os
-from shutil import copy
+from typing import List, Tuple, Union
+import dateutil
+from dateutil import parser
 
 import cf_xarray as cfxr
+import cftime
 import geopandas as gpd
-import h5py
 import importlib_metadata
 import julian
 import netCDF4 as nc
 import numpy as np
 import pandas as pd
 import xarray as xr
+import xarray.coding.times
 from shapely.geometry import Point
 from shapely.ops import transform
 
 from podaac.subsetter import dimension_cleanup as dc
 from podaac.subsetter import xarray_enhancements as xre
+from podaac.subsetter.group_handling import GROUP_DELIM, transform_grouped_dataset, recombine_grouped_datasets, \
+    h5file_transform
 
-GROUP_DELIM = '__'
 SERVICE_NAME = 'l2ss-py'
 
 
-def apply_scale_offset(scale, offset, value):
+def apply_scale_offset(scale: float, offset: float, value: float) -> float:
     """Apply scale and offset to the given value"""
     return (value + offset) / scale
 
 
-def remove_scale_offset(value, scale, offset):
+def remove_scale_offset(value: float, scale: float, offset: float) -> float:
     """Remove scale and offset from the given value"""
     return (value * scale) - offset
 
 
-def convert_bound(bound, coord_max, coord_var):
+def convert_bound(bound: np.ndarray, coord_max: int, coord_var: xr.DataArray) -> np.ndarray:
     """
-    This function will return a converted bound which which matches the
+    This function will return a converted bound, which matches the
     range of the given input file.
 
     Parameters
@@ -124,7 +128,7 @@ def convert_bound(bound, coord_max, coord_var):
     return apply_scale_offset(scale, offset, bound)
 
 
-def convert_bbox(bbox, dataset, lat_var_name, lon_var_name):
+def convert_bbox(bbox: np.ndarray, dataset: xr.Dataset, lat_var_name: str, lon_var_name: str) -> np.ndarray:
     """
     This function will return a converted bbox which matches the range
     of the given input file. This will convert both the latitude and
@@ -157,8 +161,8 @@ def convert_bbox(bbox, dataset, lat_var_name, lon_var_name):
                      convert_bound(bbox[1], 180, dataset[lat_var_name])])
 
 
-def set_json_history(dataset, cut, file_to_subset, bbox=None, shapefile=None,
-                     origin_source=None):
+def set_json_history(dataset: xr.Dataset, cut: bool, file_to_subset: str,
+                     bbox: np.ndarray = None, shapefile: str = None, origin_source=None) -> None:
     """
     Set the 'json_history' metadata header of the granule to reflect the
     current version of the subsetter, as well as the parameters used
@@ -169,15 +173,15 @@ def set_json_history(dataset, cut, file_to_subset, bbox=None, shapefile=None,
     ----------
     dataset : xarray.Dataset
         The dataset to change the header of
-    bbox : np.ndarray
-        The requested bounding box
-    file_to_subset : string
-        The filepath of the file which was used to subset
     cut : boolean
         True to cut the scanline
+    file_to_subset : string
+        The filepath of the file which was used to subset
+    bbox : np.ndarray
+        The requested bounding box
     shapefile : str
         Name of the shapefile to include in the version history
-
+    TODO: add docstring and type hint for `origin_source` parameter.
     """
 
     params = f'cut={cut}'
@@ -209,7 +213,7 @@ def set_json_history(dataset, cut, file_to_subset, bbox=None, shapefile=None,
     dataset.attrs['history_json'] = json.dumps(history_json)
 
 
-def set_version_history(dataset, cut, bbox=None, shapefile=None):
+def set_version_history(dataset: xr.Dataset, cut: bool, bbox: np.ndarray = None, shapefile: str = None) -> None:
     """
     Set the 'history' metadata header of the granule to reflect the
     current version of the subsetter, as well as the parameters used
@@ -222,10 +226,10 @@ def set_version_history(dataset, cut, bbox=None, shapefile=None):
     ----------
     dataset : xarray.Dataset
         The dataset to change the header of
-    bbox : np.ndarray
-        The requested bounding box
     cut : boolean
         True to cut the scanline
+    bbox : np.ndarray
+        The requested bounding box
     shapefile : str
         Name of the shapefile to include in the version history
 
@@ -244,7 +248,7 @@ def set_version_history(dataset, cut, bbox=None, shapefile=None):
     dataset.attrs['history'] = history.strip()
 
 
-def calculate_chunks(dataset):
+def calculate_chunks(dataset: xr.Dataset) -> dict:
     """
     For the given dataset, calculate if the size on any dimension is
     worth chunking. Any dimension larger than 4000 will be chunked. This
@@ -272,7 +276,7 @@ def calculate_chunks(dataset):
     return chunk
 
 
-def find_matching_coords(dataset, match_list):
+def find_matching_coords(dataset: xr.Dataset, match_list: List[str]) -> List[str]:
     """
     As a backup for finding a coordinate var, look at the 'coordinates'
     metadata attribute of all data vars in the granule. Return any
@@ -312,7 +316,7 @@ def find_matching_coords(dataset, match_list):
     return match_coord_vars
 
 
-def compute_coordinate_variable_names(dataset):
+def compute_coordinate_variable_names(dataset: xr.Dataset) -> Tuple[Union[List[str], str], Union[List[str], str]]:
     """
     Given a dataset, determine the coordinate variable from a list
     of options
@@ -325,7 +329,7 @@ def compute_coordinate_variable_names(dataset):
     Returns
     -------
     tuple, str
-        Tuple of strings, where the first element is the lat coordinate
+        Tuple of strings (or list of strings), where the first element is the lat coordinate
         name and the second element is the lon coordinate name
     """
 
@@ -369,7 +373,7 @@ def compute_coordinate_variable_names(dataset):
     return lat_coord_names, lon_coord_names
 
 
-def is_360(lon_var, scale, offset):
+def is_360(lon_var: xr.DataArray, scale: float, offset: float) -> bool:
     """
     Determine if given dataset is a '360' dataset or not.
 
@@ -406,7 +410,7 @@ def is_360(lon_var, scale, offset):
     return False
 
 
-def get_spatial_bounds(dataset, lat_var_names, lon_var_names):
+def get_spatial_bounds(dataset: xr.Dataset, lat_var_names: str, lon_var_names: str) -> Union[np.ndarray, None]:
     """
     Get the spatial bounds for this dataset. These values are masked
     and scaled.
@@ -415,9 +419,9 @@ def get_spatial_bounds(dataset, lat_var_names, lon_var_names):
     ----------
     dataset : xr.Dataset
         Dataset to retrieve spatial bounds for
-    lat_var_name : str
+    lat_var_names : str
         Name of the lat variable
-    lon_var_name : str
+    lon_var_names : str
         Name of the lon variable
 
     Returns
@@ -480,7 +484,7 @@ def get_spatial_bounds(dataset, lat_var_names, lon_var_names):
     return np.array([[min_lon, max_lon], [min_lat, max_lat]])
 
 
-def compute_time_variable_name(dataset, lat_var):
+def compute_time_variable_name(dataset: xr.Dataset, lat_var: xr.Variable) -> str:
     """
     Try to determine the name of the 'time' variable. This is done as
     follows:
@@ -490,7 +494,7 @@ def compute_time_variable_name(dataset, lat_var):
 
     Parameters
     ----------
-    dataset : xr.Dataset:
+    dataset : xr.Dataset
         xarray dataset to find time variable from
     lat_var : xr.Variable
         Lat variable for this dataset
@@ -530,7 +534,7 @@ def compute_time_variable_name(dataset, lat_var):
     raise ValueError('Unable to determine time variable')
 
 
-def compute_utc_name(dataset):
+def compute_utc_name(dataset: xr.Dataset) -> Union[str, None]:
     """
     Get the name of the utc variable if it is there to determine origine time
     """
@@ -541,7 +545,7 @@ def compute_utc_name(dataset):
     return None
 
 
-def get_time_epoch_var(dataset, time_var_name):
+def get_time_epoch_var(dataset: xr.Dataset, time_var_name: str) -> str:
     """
     Get the name of the epoch time var. This is only needed in the case
     where there is a single time var (of size 1) that contains the time
@@ -579,7 +583,7 @@ def get_time_epoch_var(dataset, time_var_name):
     return epoch_var_name
 
 
-def is_time_mjd(dataset, time_var_name):
+def is_time_mjd(dataset: xr.Dataset, time_var_name: str) -> bool:
     """
     Check to see if the time format is a time delta from a modified julian date.
 
@@ -603,7 +607,7 @@ def is_time_mjd(dataset, time_var_name):
     return False
 
 
-def translate_timestamp(str_timestamp):
+def translate_timestamp(str_timestamp: str) -> datetime.datetime:
     """
     Translate timestamp to datetime object
 
@@ -633,7 +637,7 @@ def translate_timestamp(str_timestamp):
     return datetime.datetime.fromisoformat(str_timestamp)
 
 
-def datetime_from_mjd(dataset, time_var_name):
+def datetime_from_mjd(dataset: xr.Dataset, time_var_name: str) -> Union[datetime.datetime, None]:
     """
     Translate the modified julian date from the long name in the time attribute.
 
@@ -665,7 +669,8 @@ def datetime_from_mjd(dataset, time_var_name):
     return None
 
 
-def build_temporal_cond(min_time, max_time, dataset, time_var_name):
+def build_temporal_cond(min_time: str, max_time: str, dataset: xr.Dataset, time_var_name: str
+                        ) -> Union[np.ndarray, bool]:
     """
     Build the temporal condition used in the xarray 'where' call which
     drops data not in the given bounds. If the data in the time var is
@@ -725,8 +730,15 @@ def build_temporal_cond(min_time, max_time, dataset, time_var_name):
     return temporal_cond
 
 
-def subset_with_bbox(dataset, lat_var_names, lon_var_names, time_var_names, variables=None, bbox=None, cut=True,
-                     min_time=None, max_time=None):
+def subset_with_bbox(dataset: xr.Dataset,
+                     lat_var_names: list,
+                     lon_var_names: list,
+                     time_var_names: list,
+                     variables=None,
+                     bbox: np.ndarray = None,
+                     cut: bool = True,
+                     min_time: str = None,
+                     max_time: str = None) -> np.ndarray:
     """
     Subset an xarray Dataset using a spatial bounding box.
 
@@ -748,11 +760,13 @@ def subset_with_bbox(dataset, lat_var_names, lon_var_names, time_var_names, vari
         ISO timestamp of min temporal bound
     max_time : str
         ISO timestamp of max temporal bound
+    TODO: add docstring and type hint for `variables` parameter.
 
     Returns
     -------
     np.array
         Spatial bounds of Dataset after subset operation
+    TODO - fix this docstring type and the type hint to match code (currently returning a list[xr.Dataset])
     """
     lon_bounds, lat_bounds = convert_bbox(bbox, dataset, lat_var_names[0], lon_var_names[0])
     # condition should be 'or' instead of 'and' when bbox lon_min > lon_max
@@ -809,7 +823,12 @@ def subset_with_bbox(dataset, lat_var_names, lon_var_names, time_var_names, vari
     return datasets
 
 
-def subset_with_shapefile(dataset, lat_var_name, lon_var_name, shapefile, cut, chunks):
+def subset_with_shapefile(dataset: xr.Dataset,
+                          lat_var_name: str,
+                          lon_var_name: str,
+                          shapefile: str,
+                          cut: bool,
+                          chunks) -> np.ndarray:
     """
     Subset an xarray Dataset using a shapefile
 
@@ -825,10 +844,13 @@ def subset_with_shapefile(dataset, lat_var_name, lon_var_name, shapefile, cut, c
         Absolute path to the shapefile used to subset the given dataset
     cut : bool
         True if scanline should be cut.
+    TODO: add docstring and type hint for `chunks` parameter.
+
     Returns
     -------
     np.array
         Spatial bounds of Dataset after shapefile subset operation
+    TODO - fix this docstring type and the type hint to match code (currently returning a xr.Dataset)
     """
     shapefile_df = gpd.read_file(shapefile)
 
@@ -871,230 +893,10 @@ def subset_with_shapefile(dataset, lat_var_name, lon_var_name, shapefile, cut, c
     return xre.where(dataset, boolean_mask, cut)
 
 
-def transform_grouped_dataset(nc_dataset, file_to_subset):
-    """
-    Transform a netCDF4 Dataset that has groups to an xarray compatible
-    dataset. xarray does not work with groups, so this transformation
-    will flatten the variables in the dataset and use the group path as
-    the new variable name. For example, data_01 > km > sst would become
-    'data_01__km__sst', where GROUP_DELIM is __.
-
-    This same pattern is applied to dimensions, which are located under
-    the appropriate group. They are renamed and placed in the root
-    group.
-
-    Parameters
-    ----------
-    nc_dataset : nc.Dataset
-        netCDF4 Dataset that contains groups
-
-    Returns
-    -------
-    nc.Dataset
-        netCDF4 Dataset that does not contain groups and that has been
-        flattened.
-    """
-
-    # Close the existing read-only dataset and reopen in append mode
-    nc_dataset.close()
-    nc_dataset = nc.Dataset(file_to_subset, 'r+')
-
-    dimensions = {}
-
-    def walk(group_node, path):
-        for key, item in group_node.items():
-            group_path = f'{path}{GROUP_DELIM}{key}'
-
-            # If there are variables in this group, copy to root group
-            # and then delete from current group
-            if item.variables:
-                # Copy variables to root group with new name
-                for var_name, var in item.variables.items():
-                    var_group_name = f'{group_path}{GROUP_DELIM}{var_name}'
-                    nc_dataset.variables[var_group_name] = var
-                # Delete variables
-                var_names = list(item.variables.keys())
-                for var_name in var_names:
-                    del item.variables[var_name]
-
-            if item.dimensions:
-                dims = list(item.dimensions.keys())
-                for dim_name in dims:
-                    new_dim_name = f'{group_path.replace("/", GROUP_DELIM)}{GROUP_DELIM}{dim_name}'
-                    item.dimensions[new_dim_name] = item.dimensions[dim_name]
-                    dimensions[new_dim_name] = item.dimensions[dim_name]
-                    item.renameDimension(dim_name, new_dim_name)
-
-            # If there are subgroups in this group, call this function
-            # again on that group.
-            if item.groups:
-                walk(item.groups, group_path)
-
-        # Delete non-root groups
-        group_names = list(group_node.keys())
-        for group_name in group_names:
-            del group_node[group_name]
-
-    for var_name in list(nc_dataset.variables.keys()):
-        new_var_name = f'{GROUP_DELIM}{var_name}'
-        nc_dataset.variables[new_var_name] = nc_dataset.variables[var_name]
-        del nc_dataset.variables[var_name]
-
-    walk(nc_dataset.groups, '')
-
-    # Update the dimensions of the dataset in the root group
-    nc_dataset.dimensions.update(dimensions)
-
-    return nc_dataset
-
-
-def recombine_grouped_datasets(datasets, output_file, start_date):  # pylint: disable=too-many-branches
-    """
-    Given a list of xarray datasets, combine those datasets into a
-    single netCDF4 Dataset and write to the disk. Each dataset has been
-    transformed using its group path and needs to be un-transformed and
-    placed in the appropriate group.
-
-    Parameters
-    ----------
-    datasets : list (xr.Dataset)
-        List of xarray datasets to be combined
-    output_file : str
-        Name of the output file to write the resulting NetCDF file to.
-    """
-
-    base_dataset = nc.Dataset(output_file, mode='w')
-
-    for dataset in datasets:
-        group_lst = []
-        for var_name in dataset.variables.keys():  # need logic if there is data in the top level not in a group
-            group_lst.append('/'.join(var_name.split(GROUP_DELIM)[:-1]))
-        group_lst = ['/' if group == '' else group for group in group_lst]
-        groups = set(group_lst)
-        for group in groups:
-            base_dataset.createGroup(group)
-
-        for dim_name in list(dataset.dims.keys()):
-            new_dim_name = dim_name.split(GROUP_DELIM)[-1]
-            dim_group = _get_nested_group(base_dataset, dim_name)
-            dim_group.createDimension(new_dim_name, dataset.dims[dim_name])
-
-        # Rename variables
-        _rename_variables(dataset, base_dataset, start_date)
-
-    # Remove group vars from base dataset
-    for var_name in list(base_dataset.variables.keys()):
-        if GROUP_DELIM in var_name:
-            del base_dataset.variables[var_name]
-
-    # Remove group dims from base dataset
-    for dim_name in list(base_dataset.dimensions.keys()):
-        if GROUP_DELIM in dim_name:
-            del base_dataset.dimensions[dim_name]
-
-    # Copy global attributes
-    base_dataset.setncatts(datasets[0].attrs)
-    # Write and close
-    base_dataset.close()
-
-
-def _get_nested_group(dataset, group_path):
-    nested_group = dataset
-    for group in group_path.strip(GROUP_DELIM).split(GROUP_DELIM)[:-1]:
-        nested_group = nested_group.groups[group]
-    return nested_group
-
-
-def _rename_variables(dataset, base_dataset, start_date):
-    for var_name in list(dataset.variables.keys()):
-        new_var_name = var_name.split(GROUP_DELIM)[-1]
-        var_group = _get_nested_group(base_dataset, var_name)
-        variable = dataset.variables[var_name]
-        var_dims = [x.split(GROUP_DELIM)[-1] for x in dataset.variables[var_name].dims]
-        if np.issubdtype(
-                dataset.variables[var_name].dtype, np.dtype(np.datetime64)
-        ) or np.issubdtype(
-            dataset.variables[var_name].dtype, np.dtype(np.timedelta64)
-        ):
-            if start_date:
-                dataset.variables[var_name].values = (dataset.variables[var_name].values - np.datetime64(start_date))/np.timedelta64(1, 's')
-                variable = dataset.variables[var_name]
-            else:
-                cf_dt_coder = xr.coding.times.CFDatetimeCoder()
-                encoded_var = cf_dt_coder.encode(dataset.variables[var_name])
-                variable = encoded_var
-
-        var_attrs = variable.attrs
-        fill_value = var_attrs.get('_FillValue')
-        var_attrs.pop('_FillValue', None)
-        comp_args = {"zlib": True, "complevel": 1}
-
-        if variable.dtype == object:
-            var_group.createVariable(new_var_name, 'S1', var_dims, fill_value=fill_value, **comp_args)
-        elif variable.dtype == 'timedelta64[ns]':
-            var_group.createVariable(new_var_name, 'i4', var_dims, fill_value=fill_value, **comp_args)
-        else:
-            var_group.createVariable(new_var_name, variable.dtype, var_dims, fill_value=fill_value, **comp_args)
-
-        # Copy attributes
-        var_group.variables[new_var_name].setncatts(var_attrs)
-
-        # Copy data
-        var_group.variables[new_var_name].set_auto_maskandscale(False)
-        var_group.variables[new_var_name][:] = variable.data
-
-
-def h5file_transform(finput):
-    """
-    Transform a h5py  Dataset that has groups to an xarray compatible
-    dataset. xarray does not work with groups, so this transformation
-    will flatten the variables in the dataset and use the group path as
-    the new variable name. For example, data_01 > km > sst would become
-    'data_01__km__sst', where GROUP_DELIM is __.
-
-    Returns
-    -------
-    nc.Dataset
-        netCDF4 Dataset that does not contain groups and that has been
-        flattened.
-    """
-    data_new = h5py.File(finput, 'r+')
-    del_group_list = list(data_new.keys())
-    has_groups = bool(data_new['/'])
-
-    def walk_h5py(data_new, group):
-        # flattens h5py file
-        for key, item in data_new[group].items():
-            group_path = f'{group}{key}'
-            if isinstance(item, h5py.Dataset):
-                new_var_name = group_path.replace('/', '__')
-
-                data_new[new_var_name] = data_new[group_path]
-                del data_new[group_path]
-
-            elif isinstance(item, h5py.Group):
-                if len(list(item.keys())) == 0:
-                    new_group_name = group_path.replace('/', '__')
-                    data_new[new_group_name] = data_new[group_path]
-
-                walk_h5py(data_new, data_new[group_path].name + '/')
-
-    walk_h5py(data_new, data_new.name)
-
-    for del_group in del_group_list:
-        del data_new[del_group]
-
-    finputnc = '.'.join(finput.split('.')[:-1]) + '.nc'
-
-    data_new.close()  # close the h5py dataset
-    copy(finput, finputnc)  # copy to a nc file
-
-    nc_dataset = nc.Dataset(finputnc, mode='r')
-
-    return nc_dataset, has_groups
-
-
-def get_coordinate_variable_names(dataset, lat_var_names=None, lon_var_names=None, time_var_names=None):
+def get_coordinate_variable_names(dataset: xr.Dataset,
+                                  lat_var_names: list = None,
+                                  lon_var_names: list = None,
+                                  time_var_names: list = None):
     """
     Retrieve coordinate variables for this dataset. If coordinate
     variables are provided, use those, Otherwise, attempt to determine
@@ -1111,6 +913,10 @@ def get_coordinate_variable_names(dataset, lat_var_names=None, lon_var_names=Non
         List of longitude coordinate variables.
     time_var_names : list
         List of time coordinate variables.
+
+    Returns
+    -------
+    TODO: add return type docstring and type hint.
     """
 
     if not lat_var_names or not lon_var_names:
@@ -1127,9 +933,19 @@ def get_coordinate_variable_names(dataset, lat_var_names=None, lon_var_names=Non
     return lat_var_names, lon_var_names, time_var_names
 
 
-def convert_to_datetime(dataset, time_vars):
+def convert_to_datetime(dataset: xr.Dataset, time_vars: list) -> Tuple[xr.Dataset, datetime.datetime]:
     """
-    converts the time variable to datetime if xarray doesn't decode times
+    Converts the time variable to datetime if xarray doesn't decode times
+
+    Parameters
+    ----------
+    dataset : xr.Dataset
+    time_vars : list
+
+    Returns
+    -------
+    xr.Dataset
+    datetime.datetime
     """
     for var in time_vars:
         start_date = datetime.datetime.strptime("1993-01-01T00:00:00.00", "%Y-%m-%dT%H:%M:%S.%f") if any('TAI93' in str(dataset[var].attrs[attribute_name])
@@ -1154,10 +970,56 @@ def convert_to_datetime(dataset, time_vars):
     return dataset, start_date
 
 
-def subset(file_to_subset, bbox, output_file, variables=None,
+def open_as_nc_dataset(filepath: str) -> Tuple[nc.Dataset, list, bool]:
+    """Open netcdf file, and flatten groups if they exist."""
+    file_extension = filepath.split('.')[-1]
+
+    if file_extension == 'he5':
+        nc_dataset, has_groups = h5file_transform(filepath)
+    else:
+        # Open dataset with netCDF4 first, so we can get group info
+        nc_dataset = nc.Dataset(filepath, mode='r')
+        has_groups = bool(nc_dataset.groups)
+
+        # If dataset has groups, transform to work with xarray
+        if has_groups:
+            nc_dataset = transform_grouped_dataset(nc_dataset, filepath)
+
+    nc_dataset, rename_vars = dc.remove_duplicate_dims(nc_dataset)
+
+    return nc_dataset, rename_vars, has_groups
+
+
+def override_decode_cf_datetime() -> None:
+    """
+    WARNING !!! REMOVE AT EARLIEST XARRAY FIX, this is a override to xarray override_decode_cf_datetime function.
+    xarray has problems decoding time units with format `seconds since 2000-1-1 0:0:0 0`, this solves by testing
+    the unit to see if its parsable, if it is use original function, if not format unit into a parsable format.
+
+    https://github.com/pydata/xarray/issues/7210
+    """
+
+    orig_decode_cf_datetime = xarray.coding.times.decode_cf_datetime
+
+    def decode_cf_datetime(num_dates, units, calendar=None, use_cftime=None):
+        try:
+            parser.parse(units.split('since')[-1])
+            return orig_decode_cf_datetime(num_dates, units, calendar, use_cftime)
+        except dateutil.parser.ParserError:
+            reference_time = cftime.num2date(0, units, calendar)
+            units = f"{units.split('since')[0]} since {reference_time}"
+            return orig_decode_cf_datetime(num_dates, units, calendar, use_cftime)
+
+    xarray.coding.times.decode_cf_datetime = decode_cf_datetime
+
+
+def subset(file_to_subset: str, bbox: np.ndarray, output_file: str,
+           variables: Union[List[str], str, None] = (),
            # pylint: disable=too-many-branches, disable=too-many-statements
-           cut=True, shapefile=None, min_time=None, max_time=None, origin_source=None,
-           lat_var_names=None, lon_var_names=None, time_var_names=None):
+           cut: bool = True, shapefile: str = None, min_time: str = None, max_time: str = None,
+           origin_source: str = None,
+           lat_var_names: List[str] = (), lon_var_names: List[str] = (), time_var_names: List[str] = ()
+           ) -> Union[np.ndarray, None]:
     """
     Subset a given NetCDF file given a bounding box
 
@@ -1165,16 +1027,14 @@ def subset(file_to_subset, bbox, output_file, variables=None,
     ----------
     file_to_subset : string
         The location of the file which will be subset
-    output_file : string
-        The file path for the output of the subsetting operation.
     bbox : np.ndarray
         The chosen bounding box. This is a tuple of tuples formatted
         as such: ((west, east), (south, north)). The assumption is that
         the valid range is ((-180, 180), (-90, 90)). This will be
         transformed as appropriate if the actual longitude range is
         0-360.
-    shapefile : str
-        Name of local shapefile used to subset given file.
+    output_file : string
+        The file path for the output of the subsetting operation.
     variables : list, str, optional
         List of variables to include in the resulting data file.
         NOTE: This will remove ALL variables which are not included
@@ -1182,6 +1042,8 @@ def subset(file_to_subset, bbox, output_file, variables=None,
     cut : boolean
         True if the scanline should be cut, False if the scanline should
         not be cut. Defaults to True.
+    shapefile : str
+        Name of local shapefile used to subset given file.
     min_time : str
         ISO timestamp representing the lower bound of the temporal
         subset to be performed. If this value is not provided, the
@@ -1190,6 +1052,9 @@ def subset(file_to_subset, bbox, output_file, variables=None,
         ISO timestamp representing the upper bound of the temporal
         subset to be performed. If this value is not provided, the
         granule will not be subset temporally on the upper bound.
+    origin_source : str
+        Original location or filename of data to be used in "derived from"
+        history element.
     lat_var_names : list
         List of variables that represent the latitude coordinate
         variables for this granule. This list will only contain more
@@ -1206,23 +1071,21 @@ def subset(file_to_subset, bbox, output_file, variables=None,
         than one value in the case where there are multiple groups and
         different coordinate variables for each group.
     """
-    file_extension = file_to_subset.split('.')[-1]
+    nc_dataset, rename_vars, has_groups = open_as_nc_dataset(file_to_subset)
 
-    if file_extension == 'he5':
-        nc_dataset, has_groups = h5file_transform(file_to_subset)
-    else:
-        # Open dataset with netCDF4 first, so we can get group info
-        nc_dataset = nc.Dataset(file_to_subset, mode='r')
-        has_groups = bool(nc_dataset.groups)
+    override_decode_cf_datetime()
 
-        # If dataset has groups, transform to work with xarray
-        if has_groups:
-            nc_dataset = transform_grouped_dataset(nc_dataset, file_to_subset)
-
-    nc_dataset, rename_vars = dc.remove_duplicate_dims(nc_dataset)
-
-    if variables:
-        variables = [x.replace('/', GROUP_DELIM) for x in variables]
+    if has_groups:
+        # Make sure all variables start with '/'
+        variables = ['/' + var if not var.startswith('/') else var for var in variables]
+        lat_var_names = ['/' + var if not var.startswith('/') else var for var in lat_var_names]
+        lon_var_names = ['/' + var if not var.startswith('/') else var for var in lon_var_names]
+        time_var_names = ['/' + var if not var.startswith('/') else var for var in time_var_names]
+        # Replace all '/' with GROUP_DELIM
+        variables = [var.replace('/', GROUP_DELIM) for var in variables]
+        lat_var_names = [var.replace('/', GROUP_DELIM) for var in lat_var_names]
+        lon_var_names = [var.replace('/', GROUP_DELIM) for var in lon_var_names]
+        time_var_names = [var.replace('/', GROUP_DELIM) for var in time_var_names]
 
     args = {
         'decode_coords': False,